backend/app/requirement.txt
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.32
cryptography==43.0.0
mysqlclient==2.2.4
python-dotenv==1.0.1